Rmove create page and update masternode page
